pyproject.toml
[tool.poetry]
name = "momentum-ia"
version = "0.1.0"
package-mode = false
description = ""
authors = ["SantiPandal <146788383+SantiPandal@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
fastapi = ">=0.115.12,<0.116.0"
uvicorn = ">=0.34.3,<0.35.0"
python-multipart = ">=0.0.20,<0.0.21"
twilio = ">=9.6.2,<10.0.0"
python-dotenv = ">=1.1.0,<2.0.0"
langchain = ">=0.3.25,<0.4.0"
langchain-openai = ">=0.3.19,<0.4.0"
langgraph = ">=0.2.0,<0.3.0"
supabase = ">=2.15.2,<3.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.21.0"
pytest-mock = "^3.12.0"
httpx = "^0.25.0"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
